Remove LocalWorker class
The idea behind this class is to declare Master to be a friend of
LocalWorker and get access to local_worker's private fields.

However, it doesn't work because friend declaration doesn't apply
to the fields inherited from WorkerBase.

So the solution is to remove LocalWorker class entirely and use
WorkerBase class in place of it.